requirements.txt
# LLM Alert Enrichment Service requirements
python-dotenv
pydantic
requests
opensearch-py
